refactor: useBoards 추가 (#60)
