+++ src/components/board/Editor.tsx
@@ -1,16 +1,9 @@
 import { getPost, insertNewPost, updatePost } from '@/apis/boards';
-import { toastSuccess, toastWarn } from '@/libs/toastifyAlert';
+import { useBoards } from '@/hooks/useBoards';
+import { toastWarn } from '@/libs/toastifyAlert';
 import { RootState } from '@/redux/config/configStore';
-import {
-  Button,
-  Input,
-  Textarea,
-  Select,
-  SelectItem,
-  Spacer,
-  Divider,
-} from '@nextui-org/react';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { Button, Input, Textarea, Spacer, Divider } from '@nextui-org/react';
+import { useQuery } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
@@ -25,6 +18,7 @@ interface Props {
 const Editor = ({ isEdit }: Props) => {
   const { userId } = useSelector((state: RootState) => state.auth);
   const router = useRouter();
+  const { insertPost, updatePost } = useBoards();
   const boardId: string = router.query.boardId as string;
   const [inputs, setInputs] = useState({
     title: '',
@@ -37,25 +31,6 @@ const Editor = ({ isEdit }: Props) => {
     queryKey: ['posts', boardId],
     queryFn: () => getPost(boardId),
     enabled: !!boardId,
-  });
-
-  const queryClient = useQueryClient();
-  const { mutate: addPostMutate } = useMutation({
-    mutationFn: insertNewPost,
-    onSuccess: () => {
-      router.push('/board');
-      toastSuccess('등록되었습니다!');
-      queryClient.invalidateQueries({ queryKey: ['posts'] });
-    },
-  });
-
-  const { mutate: updatePostMutate } = useMutation({
-    mutationFn: updatePost,
-    onSuccess: () => {
-      router.push('/board');
-      toastSuccess('수정되었습니다!');
-      queryClient.invalidateQueries({ queryKey: ['posts'] });
-    },
   });
 
   const inputChange = (
@@ -106,7 +81,7 @@ const Editor = ({ isEdit }: Props) => {
       confirmButtonColor: '#FFD029',
     }).then((result) => {
       if (result.isConfirmed) {
-        addPostMutate(formData);
+        insertPost(formData);
       }
     });
   };
@@ -124,7 +99,7 @@ const Editor = ({ isEdit }: Props) => {
       confirmButtonColor: '#FFD029',
     }).then((result) => {
       if (result.isConfirmed) {
-        updatePostMutate({ boardId, editValue: formData });
+        updatePost({ boardId, editValue: formData });
       }
     });
   };
